refactor: use named exports for app components
